Fix HD gameover, merge succubus functions into succubus.c
